Restore environment variable override behaviour
Sometime after v0.14 a syntatic change was made to how env vars
and passwd config settings were merged which inverted the precedence
behaviour and prevented env vars from overriding config settings.
This commit restores the original precedence order. Without this change,
the value of an env var is only used for a setting which is not
specified as a parameter to the `BaseSettings` constructor.

This commit also changes the relevant test behaviour, to make it expect
the env var rather than the passed parameter.

Fixes #340