Allow for passing keyword arguments to from_orm:
You can already use Field(alias='whatever') on the model to allow for
converting from ORM models that might be named differently, but this
means that the model knows about the existence of a database (or at
least it's columns), which seems like a leaky abstraction.

With this modification though, the database will be the one to handle
the naming conflict, which keeps the model more succinct.