feat: Accept `extra` kwarg in `BaseModel.copy`
Add a new keyword argument (`extra`) to `BaseModel.copy` that provides behavior
similar to that of the `BaseConfig` attribute. When `extra == Extra.ignore`, any
keys in the `update` dict are ignored. Similarly, when `extra == Extra.forbid`
a `CopyError` is raised similar to the `ValidationError` that would be raised if
the constructor were called with the same dict.