Fix TypeError for GenericModel with Callable param
Introduced in 1.10.2, a TypeError would be raised upon creation of a
GenericModel class that used a Callable type parameter. This would
happen when `typing.get_args` returned a list for the first type
agruments in a Callable and pydantic would try to use the value as a
dictionary key. To avoid the issue, we convert the list to a tuple
before using it as a key.

The possible approach of modifying pydantic's `get_args` function to
return a tuple instead of a list didn't work out because the return
values are used in more places, some of which expect the list and not a
tuple.

Fixes #4551