Fix racy doctests
On very fast machines (in my case, a 7950X3D with gen4 storage) it's possible that the script will start running the tests before importlib figures out the files are there, and get a spurious `ModuleNotFoundError` instead. Explicitly flushing the cache before importing things seems to help it work consistently, at the cost of an extra 0.1s or so of runtime.