Mention PyObject (v1) moving to ImportString (v2)
It seems like the equivalent of PyObject from v1
is the ImportString type in v2,
but there's no mention of it in the migration docs.

There's a "known limitation" note in the ImportString docs,
not sure if I should mention it in the migration doc.

BTW the doc also states:
> This is actively being worked on.

can this be tracked somewhere?

Also, not sure if I should update `pydantic/_migration.py` too 🤔.