JSON schema: fix extra parameter handling
The core schema for typed dicts has an "extra_behavior" key for checking
whether additional unspecified keys are allowed, but the JSON schema
generator ignores this and only checks the configuration dict.  Have the
JSON schema generator check both places, and prefer the value set in the
core schema over the one in the config dictionary if both are present.

Fixes #7809.