Improve class detection when walking up frames
Still needs additional tests, I think we can hit edge cases where
detection still doesn't work as expected.
By checking if we find the `class` statement when walking up frames,
we are able to make it work with dataclasses/TypedDicts.
Using a regexp match was required, or it would lead to false positives,
and we can't just use startswith as there can be an arbitrary number of
spaces between `class` and the name of that class.