Modify parse_raw() and parse_file() not to emit spurious warnings
Modify the `parse_raw()` and `parse_file()` to ignore deprecation
warnings from their implementation details.  The methods themselves
are deprecated, and that should be the only deprecation warning
interesting to the user.  This also fixes test failures with pytest-8
that no longer ignores warnings that did not match `pytest.warns()`.