Update dependencies across requirements files
Pin the following packages to play nice with Python 3.12.
  * Flake8
  * Hypothesis
  * Cython

Pin mike to a tag instead of chasing the HEAD of the default branch.

Add tomli to make mypy complain less when running pre-commit hooks.

Builds require distutils, but that module was removed from the stdlib
in 3.12. Add setuptools which provides a distutils wrapper.

Signed-off-by: Vincent Fazio <vfazio@gmail.com>