setup: remove upper bound from python_requires
A cap was recently added to denote the upper bound of python support for
pydantic v1.

Further research, however, has determined that `python_requires` was not
developed with an upper bound in mind [0] and may actually cause some
dependency solvers to silently fall back to older versions where there
is no cap, and thus presumed support for the user's python.

For now, remove the upper bound.

[0]: https://discuss.python.org/t/requires-python-upper-limits/12663

Signed-off-by: Vincent Fazio <vfazio@gmail.com>