Add note about Pipeline API recommended usage
I think it's important to make this clarification.
Without it I fear people may get excited about the shiny new thing or go too wild with it just because of how much it can do, which IMO would be a mistake. I view this as an advanced feature, like descriptors or something: when you need it it's awesome it's there and saves you a lot of pain, but use it everywhere for everything and you'll end up with a spaghetti codebase that's hard to navigate and reason about.