Allow validators to customize validation JSON schema
The annotated validators now have a `input_type` argument.
Users can provide a type hint reflecting the allowed type
for the validator. Unlike serializers where the type hint
can be inferred from the return type annotation of the
function, we don't allow this pattern in our case as we want
to enforce the validator functions to check for any type
(by using `typing.Any`).