Handle `nullable` schemas with `serialization` schema available during JSON Schema generation (#10132)
It might be that the `serialization` schema has a specific `when_used`
mode set to `unless-none` or `json-unless-none`. In that case,
if the base core schema is of type `nullable`, we need to account
for this as the `serialization` schema might not be used.

This also fixes an issue where `unless-none` was not taken into
account when trying to figure out if the default value should be
passed into the plain serializer function to get the JSON Schema
default.