Refactor how `field_validator` decorators are applied during core schema generation
This reduces code duplication, as the only source of truth is now
defined in the `__get_pydantic_core_schema__` methods.
The `apply_validators` function is kept for deprecated
decorators but will be removed at some point.

Additionally, this allows `input_type` to be taken into account,
without having to duplicate code even more.