Update and clean up several aspects of documentation tooling (#10203)
Remove `mkdocs-simple-hooks` dependency, this is now available in MkDocs
Remove `mkdocs-embed-external-markdown` and apply same logic in a hook
instead. The extension was parsing every markdown file as Jinja2
templates, which had limitations with some syntax. There's also
a small performance improvement in not doing so.
Update `griffe` to the latest version, fixing breaking changes
to extension, and remove dependency to `backports-strenum`.
Do not unnecessarily depend on a git version of `mike`