Add support for dict with enum keys typing
JSON schema draft 2020-12 defines a propertyNames key for dict/object
schemas which can contain any valid JSON schema that defines to what any
property name of the object instance needs to validate to.

This can be mapped to a pydantic/python typing like dict[AnyEnum, str]
where AnyEnum is an Enum type with defined values.

propertyNames supports other validation schemas as well, e.g. a pattern.
However, this is not covered with this commit.

Fixes #4393