Fix attributes format of code blocks in documentation
It was currently using a custom format to properly render in IDEs,
but the ```python {attrs} format also works and IDEs do understand
it this way.

We had some code in our documentation plugin to revert the changes,
but it only applied to actual markdown documentation and not docstrings
in source code.

Also use `python` as a language specifier consistently.