Rework `create_model` field definitions format
If the value is a two-tuple, the first element is the type and
the second element is the assigned value (either a default or
a `Field()` function). This isn't breaking as this was already
allowed before.

We now also allow a single element to be passed, which will be
the type. This also isn't breaking because:
- We previously allowed a single element to be passed if it was
  an annotated type. We then extracted the first metadata element
  and enforced it to be a `Field()` function. With these changes,
  we now just leave the annotated type as is, so an arbitrary amount
  of metadata can now be specified.
- If the type wasn't annotated, we were raising an error.

Private fields are now allowed.

Redundant tests were merged/removed. Documentation was updated.

The `__slots__` parameter was removed. It was deprecated (with
a wrong warning class, but still) since 1.10, was not doing anything,
so I don't expect anyone to be using it.