Make sure the type reference is removed from the seen references (#11143)
When parametrizing generic models, the `generic_recursion_self_type`
context manager is entered, and is used to avoid infinite recursions
if the same generic model happens to be parametrized again with the
same args during the first parametrization.

However, upon exiting the context manager (and thus when the parametrized
model is fully created), we forgot to remove the type ref from the set.
This happened only if we were already in the process of parametrizing
another model, as otherwise the `_generic_recursion_cache` would
be reset (see the `if token` condition).

In theory, this couldn't cause issues because parametrized models are
cached, and the cache is checked *before* entering the context manager.
However, because we have custom `mro()` implementation on the `BaseModel`
metaclass, this ends up causing issues is some really specific scenarios.

Co-authored-by: kc0506 <kchong0506@gmail.com>