Reorganize `extra` configuration documentation (#11213)
The section in the concepts page was moved to the top, as
the default behavior is usually confusing for some users.

The section was also simplified to just give some insights,
and we refer to the API docs to avoid duplication. The API
documentation was updated to take examples from both the
existing concepts section and API documentation.

Add more info about support for extra data with dataclasses

Remove unused annotation comment.