Recollect fields during rebuild if collection failed
This is the biggest change in this PR, and the hardest to
reason about and the trickiest to implement.

Instead of re-evaluating annotations on a field-by-field basis,
we assume all fields have their annotation evaluated (and
consequently the `FieldInfo` attributes -- some of them
set depending on the annotation -- are correct). To do so,
we collect model fields (if necessary, thanks to the new
`__pydantic_fields_complete__` attribute) in `_model_schema()`,
and then have `_common_field_schema()` called with a fully built
`FieldInfo` instance (this mimics what is done for dataclasses).

When `model_rebuild()` is called, we also want to recollect fields
if it wasn't successful the first time. This introduces challenges,
such as keeping the model assignments to be able to recollect fields
and call `FieldInfo.from_annotated_attribute()`.