Collect model fields when rebuilding a model (#11388)
Currently, unknown forward annotations doesn't stop field collection. This introduces subtle bugs and code smells that this PR tries removing.

The first issue arising from this is that the model fields can be inaccurate when a model is not fully built yet:

```python
class Model(BaseModel):
    a: 'Annotated[Forward, Field(alias="b")]'

Model.model_fields['a'].alias
#> None, should be 'b'
```

When a model is being rebuilt, we reevaluate the annotation and recreate a F`ieldInfo` instance during the core schema generation process. Doing so introduced a code smell where we track explicitly set attributes using `_attributes_set`. This is known to be a source of bugs as we need to keep track of such explicitly set attributes when you alter the `FieldInfo` instance.

This fix introduces a new `__pydantic_fields_complete__` class property. It is true if at least one field annotation failed to evaluate during collect_model_fields(). Calling `model_rebuild()` will redo the fields collection if it wasn't successful the first time.

When an incomplete Pydantic model is encountered as an annotation (inside `GenerateSchema`), we first call `rebuild_model_fields()`; a new function that iterates over the existing (but potentially not complete) field infos, and recreate instances if the annotation needs to be evaluated again. This way, we avoid generating a schema for each field just to realize that one field still has a forward annotation.