Update fields collection logic
The source is provided, and the logic to detect final fields
with a default value is moved *after* the `FieldInfo` instance
is created.

The reason to do so is to again support more edge cases, e.g.
When the annotation was wrapped with `Annotated` (e.g.
`Annotated[Final[int], ...] = 1`. This previously wasn't considered
and as such wasn't treated as a classvar. As per the linked issue
in this PR, this led to an inconsistency, so while technically
this can be considered as a breaking change, we'll consider this as
a bug.