Provide the available definitions when applying discriminated unions
It seems that we forgot to do so, and as almost all discriminated unions
involves referenceable schemas (such as models, typed dicts, etc), the
deferred metadata key was almost always set.

This fixes some issues in the schema cleaning process, although it is
still theoretically unsound (see the added tests).