Move field-related logic out of the `GenerateSchema` class
This is where most of the work is being done. A new `update_field_from_config()`
function is defined and can be used to update the `FieldInfo` instances
from the configuration of the model/class the field belongs to.

Currently, this function will update the field instances by applying
alias and title generators.

Two reasons to move this logic out:
- Alleviate the `GenerateSchema` class, which is currently a big monster.
- Apply these updates as early as possible, when models are being created.
  This way (assuming annotations where resolved), users can safely inspect
  `FieldInfo.alias/title` even if `defer_build=True`.