Apply same logic for computed fields
This is done through the `DecoratorInfos` via a new method.
Ideally, we could also pass the `config_wrapper` in the existing
`build()` method, but this method is recursive so it isn't
straightforward. This has the unfortunate consequence of
having to create the `DecoratorInfos` in two steps/calls, which
is error prone. Perhaps in the future `build()` can be refactored
to avoid the recursive calls.