BaseModel: use `typing.Dict` for dunder annotations
There is a disclaimer at top of file that mentions using `typing.Dict` for
annotations rather than `dict` itself due to `dict` being part of the
`BaseModel` namespace (in the form of the deprecated `dict()` method). We
correct two violations of this disclaimer in this PR, which currently cause
downstream issues in packages that rely heavily on parsing of type-annotations,
e.g. mitmproxy/pdoc#792.