Improve SEO for older versions
Update some canonical links that currently point to redirects.
Search engines will discard the canonical link if it is too
different from the current one, which is the case with redirects
as it only contains the JS code to make the redirection.

Prevent indexing of the Computed fields page, which was moved
to be a page section and as such older version were still poping up.